Use HDF5 file_open from C side (and remove file_create)
