+++ src/summary.F90
@@ -38,7 +38,7 @@ contains
     call write_message("Writing summary.h5 file...", 5)
 
     ! Create a new file using default properties.
-    file_id = file_create("summary.h5")
+    file_id = file_open("summary.h5", 'w')
 
     call write_header(file_id)
     call write_nuclides(file_id)
